% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.R
\name{batch_validate}
\alias{batch_validate}
\title{Batch validate algorithms over a list of dfms}
\usage{
batch_validate(x, y, set = docvars(x, "training"), alg = NULL, pred = NULL)
}
\arguments{
\item{x}{A dfm or list of dfms to evaluate.}

\item{y}{A variable in docvars of the dfms with the class (must be logical
for now).}

\item{set}{A variable in docvars indicating membership to training or test
set.}

\item{alg}{A named list of functions containing the algorithms to be
evaluated (see example)}

\item{pred}{A list of prediction functions (most packages use predict(),
which is the default). Either length of 1 or same length as alg.}
}
\description{
Batch validate algorithms over a list of dfms
}
\examples{
\dontrun{

library(quanteda)
corp <- corpus(c(d1 = "Chinese Beijing Chinese",
                 d2 = "Chinese Chinese Shanghai",
                 d3 = "Chinese Macao",
                 d4 = "Tokyo Japan Chinese",
                 d5 = "Chinese Chinese Chinese Tokyo Japan"))

docvars(corp, "class") <- c(TRUE, TRUE, TRUE, FALSE, FALSE)
docvars(corp, "training") <- c(TRUE, TRUE, TRUE, TRUE, FALSE)

test_prep <- batch_prep(corp)

batch_validate(x = test_prep[1:3],
               y = "class",
               set = "training",
               alg = list(textmodel_nb = quanteda.textmodels::textmodel_nb,
                          textmodel_svm = quanteda.textmodels::textmodel_svm)
}
}
