% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confu_mat.R
\name{confu_mat}
\alias{confu_mat}
\title{Flat confusion matrix}
\usage{
confu_mat(pred, true, case_name = NULL, positive = TRUE)
}
\arguments{
\item{pred}{Predicted classes.}

\item{true}{True classes.}

\item{case_name}{Name added to the tibble.}

\item{positive}{Value for the positive class.}
}
\value{
A flat confusion matrix accuracy measures.
}
\description{
Flat confusion matrix
}
