% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.R
\name{val}
\alias{val}
\title{Evaluate  an algorithm with one dfm}
\usage{
val(
  x,
  y,
  set = "training",
  alg = NULL,
  pred = stats::predict,
  as_matrix = FALSE,
  positive = TRUE,
  pb = NULL,
  what = NULL,
  write_out = FALSE
)
}
\arguments{
\item{x}{A dfm to evaluate.}

\item{y}{A variable in docvars of the dfms with the class (must be logical
for now).}

\item{set}{A variable in docvars indicating membership to training or test
set.}

\item{alg}{A functions containing the algorithm to be evaluated (see example)}

\item{pred}{A function used for prediction (most packages use predict(),
which is the default).}

\item{as_matrix}{Logical. Indicating if alg needs the dfm converted to a
matrix.}

\item{positive}{value for the positive class.}

\item{pb, what}{Used to display status bar when used in batch mode.}

\item{write_out}{Should intermediate results be written to the working
directory.}
}
\value{
A confusion matrix object.
}
\description{
Evaluate  an algorithm with one dfm
}
